refactor: FF-22 renamed adapters folder
